+++ packages/core/src/index.ts
@@ -27,6 +27,6 @@ export { default as Position } from './portal/Position';
 export { default as Tooltip } from './portal/Tooltip';
 export { default as SpecialZoomLevel } from './SpecialZoomLevel';
 export { default as createStore } from './store/createStore';
-export { default as TextLayerRenderStatus } from './types/TextLayerRenderStatus';
+export { default as LayerRenderStatus } from './types/LayerRenderStatus';
 export { default as Viewer } from './Viewer';
 export { default as Worker } from './Worker';